refactor(domain): convert byte to int priority attribute on Price

+++ src/main/java/com/alanviera/ecommerce/domain/model/Price.java
@@ -14,7 +14,7 @@ public class Price {
     private final Long priceListId;
     private final LocalDateTime startDate;
     private final LocalDateTime endDate;
-    private final byte priority;
+    private final int priority;
     private final BigDecimal price;
     private final String currency;
 }
